Extract horizontal scroll into a component
