+++ frontend/src/components/header/styles.scss
@@ -1,15 +1,5 @@
-.menu-items-container::-webkit-scrollbar {
+.navigation-items-container::-webkit-scrollbar {
   display: none;
-}
-
-.translate-icon {
-  top: 50%;
-  transform: translateY(-50%);
-}
-
-.translate-icon-btm {
-  top: 50%;
-  transform: translateY(-50%) rotate(180deg);
 }
 
 @media screen and (max-width: 45em) {
