+++ frontend/src/components/projectDetail/footer.js
@@ -1,15 +1,13 @@
-import React, { useState, useEffect, useRef, Fragment } from 'react';
+import React, { useRef, Fragment } from 'react';
 import { Link } from '@reach/router';
 import { useSelector } from 'react-redux';
 import { FormattedMessage } from 'react-intl';
 
-import useForceUpdate from '../../hooks/UseForceUpdate';
-import { useWindowSize } from '../../hooks/UseWindowSize';
 import { Button } from '../button';
 import messages from './messages';
 import { ShareButton } from './shareButton';
 import { AddToFavorites } from './favorites';
-import { ChevronRightIcon } from '../svgIcons';
+import { HorizontalScroll } from '../horizontalScroll';
 
 import './styles.scss';
 
@@ -46,41 +44,7 @@ const menuItems = [
 
 export const ProjectDetailFooter = ({ className, projectId }) => {
   const userIsloggedIn = useSelector((state) => state.auth.token);
-  const [scrollLeft, setScrollLeft] = useState(0);
   const menuItemsContainerRef = useRef(null);
-  const size = useWindowSize();
-  const [, forceUpdate] = useForceUpdate();
-
-  useEffect(() => {
-    forceUpdate();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, size);
-
-  useEffect(() => {
-    const menuItemsContainer = document.querySelector('.menu-items-container');
-    menuItemsContainer.addEventListener('scroll', updateScrollLeft);
-
-    return () => {
-      menuItemsContainer.removeEventListener('scroll', updateScrollLeft);
-    };
-  }, []);
-
-  const updateScrollLeft = (e) => {
-    setScrollLeft(e.target.scrollLeft);
-  };
-
-  const handleScroll = (direction) => {
-    let currentScroll = scrollLeft;
-    if (direction === 'right') {
-      currentScroll += 200;
-    } else {
-      currentScroll -= 200;
-    }
-    menuItemsContainerRef.current.scrollTo({
-      left: currentScroll,
-      behavior: 'smooth',
-    });
-  };
 
   return (
     <div
@@ -90,26 +54,11 @@ export const ProjectDetailFooter = ({ className, projectId }) => {
       style={{ boxShadow: '0px -1px 0px #F0EFEF, 0px 1px 0px #F0EFEF' }}
     >
       {/* TODO ADD ANCHORS */}
-      <div className="relative overflow-hidden w-60-ns pr1 mr4 dn db-l">
-        <div
-          className={`bg-white absolute left-0 rotate-180 z-1 pointer pl2 ${
-            scrollLeft > 0 ? 'db' : 'dn'
-          }`}
-          onClick={() => handleScroll('left')}
-        >
-          <ChevronRightIcon />
-        </div>
-        <div
-          className={`bg-white absolute right-0 z-1 pointer pl2 ${
-            scrollLeft <
-            menuItemsContainerRef.current?.scrollWidth - menuItemsContainerRef.current?.clientWidth
-              ? 'db'
-              : 'dn'
-          }`}
-          onClick={() => handleScroll('right')}
-        >
-          <ChevronRightIcon />
-        </div>
+      <HorizontalScroll
+        className={'w-60-ns pr1 mr4 dn db-l'}
+        menuItemsContainerRef={menuItemsContainerRef}
+        containerClass={'.menu-items-container'}
+      >
         <div ref={menuItemsContainerRef} className="menu-items-container nowrap overflow-x-auto">
           {menuItems.map((menuItem, index) => (
             <Fragment key={menuItem.href}>
@@ -120,7 +69,7 @@ export const ProjectDetailFooter = ({ className, projectId }) => {
             </Fragment>
           ))}
         </div>
-      </div>
+      </HorizontalScroll>
       <div className="flex items-center ml-auto gap-1">
         <ShareButton projectId={projectId} />
         {userIsloggedIn && <AddToFavorites projectId={projectId} />}
